add generic get_dumpfile_regs to read registers
Currently the logic for reading registers from each type of dumpfile is
separated into different functions depending on the vmcore type.
Since this code was required at multiple places, so instead of
duplicating the code, have `get_dumpfile_regs` do that, which will
call the appropriate function for reading the registers based on the
vmcore type and architecture.

This is used to implement "frame"/"up"/"down" in crash, and has been
used in later commit to fix gdb passthrough by providing it register values

Signed-off-by: Aditya Gupta <adityag@linux.ibm.com>