make cpu context change transparent to crash/gdb
Currently the gdb passthrough get back to `crash_target::fetch_registers`
to get register values, which depends on the current context in crash.

So even though `thread` command run in gdb mode or as a passthrough,
it did not change the cpu in crash, hence no context change in crash,
thereby getting register values with respect to wrong thread/cpu

Even if in gdb mode, `thread <thread_num>` command is run, it does not
change the cpu in the crash context.

Modify the current cpu in crash, when 'thread <thread_num' is run in gdb
mode or as gdb passthrough.

Signed-off-by: Aditya Gupta <adityag@linux.ibm.com>