thread: set cpu in crash on gdb "thread" change also
Even if in gdb mode, `thread <thread_num>` command is run, it does not
change the cpu in the crash context.

This causes an issue since, gdb passthrough depends on `crash_target` to
get the register values, which is dependent on the current CPU selected
in crash with `set -c <num>`. the gdb passthroughs get back to `crash_target` to get register
information, which depends on the current context in crash.

So even though `thread` command run in gdb mode or as a passthrough,
it did not change the cpu in crash, hence no context change in crash,
thereby getting register values with respect to wrong thread/cpu

Signed-off-by: Aditya Gupta <adityag@linux.ibm.com>