ppc64: fix gdb passthrough by implementing machdep->get_cpu_reg
Currently, gdb passthroughs of 'bt', 'frame', 'up', 'down', 'info
locals' don't work. This is due to gdb not knowing the register values to
unwind the stack frames

Every gdb passthrough goes through `gdb_interface`. And then, gdb expects
`crash_target::fetch_registers` to give it the register values, which is
dependent on `machdep->get_cpu_reg` to read the register values for
specific architecture.

                                      ┌────────────────────────┐
           gdb passthrough (eg. "bt") │                        │
   crash   ─────────────────────────▶│                        │
                                      │      gdb_interface     │
                                      │                        │
                                      │                        │
                                      │ ┌────────────────────┐ │
                 fetch_registers      │ │                    │ │
crash_target◀────────────────────────┼─┤        gdb         │ │
            ─────────────────────────┼▶│                    │ │
              Registers (SP,NIP, etc.)│ │                    │ │
                                      │ │                    │ │
                                      │ └────────────────────┘ │
                                      └────────────────────────┘

Implement `machdep->get_cpu_reg` on PPC64, so that crash provides the
register values to gdb to unwind stack frames properly

Also, with the goal to preserve the behaviour of gdb passthroughs in
default mode, a hook 'machdep->gdb_mode_hook' has been introduced, which
will get executed everytime when "set gdb on", and "set gdb off" are run,
and can be used by any architecture.
This hook has been used on PPC64 such that, `machdep->get_cpu_reg` is only
set on "set gdb on", and set to NULL in default mode (or on "set gdb off"),
so that behaviour of gdb passthroughs in default mode stays the same.

With these changes, on powerpc, 'bt' command output in gdb mode, will look
like this:

    gdb> bt
    #0  0xc0000000002a53e8 in crash_setup_regs (oldregs=<optimized out>, newregs=0xc00000000486f8d8) at ./arch/powerpc/include/asm/kexec.h:69
    #1  __crash_kexec (regs=<optimized out>) at kernel/kexec_core.c:974
    #2  0xc000000000168918 in panic (fmt=<optimized out>) at kernel/panic.c:358
    #3  0xc000000000b735f8 in sysrq_handle_crash (key=<optimized out>) at drivers/tty/sysrq.c:155
    #4  0xc000000000b742cc in __handle_sysrq (key=key@entry=99, check_mask=check_mask@entry=false) at drivers/tty/sysrq.c:602
    #5  0xc000000000b7506c in write_sysrq_trigger (file=<optimized out>, buf=<optimized out>, count=2, ppos=<optimized out>) at drivers/tty/sysrq.c:1163
    #6  0xc00000000069a7bc in pde_write (ppos=<optimized out>, count=<optimized out>, buf=<optimized out>, file=<optimized out>, pde=0xc000000009ed3a80) at fs/proc/inode.c:340
    #7  proc_reg_write (file=<optimized out>, buf=<optimized out>, count=<optimized out>, ppos=<optimized out>) at fs/proc/inode.c:352
    #8  0xc0000000005b3bbc in vfs_write (file=file@entry=0xc00000009dda7d00, buf=buf@entry=0xebcfc7c6040 <error: Cannot access memory at address 0xebcfc7c6040>, count=count@entry=2, pos=pos@entry=0xc00000000486fda0) at fs/read_write.c:582

instead of earlier output without this patch:

    gdb> bt
    #0  <unavailable> in ?? ()
    Backtrace stopped: previous frame identical to this frame (corrupt stack?)

Signed-off-by: Aditya Gupta <adityag@linux.ibm.com>