diskdump: add hook for additional checks on prstatus notes validity
Introduce a hook in machdep, for architectures to determine validity of
PRSTATUS notes, when reading percpu prstatus notes.

Also a default hook has been provided for all architectures other than
PowerPC64, which checks if 'crash_notes' for a given cpu is valid

On PowerPC64, since it doesn't utilise 'crash_notes' to get register
values, so no additional checks are required

TODO: Move to patch extras, after Signed-Off-By

Issue faced due to current code
===============================

Recently an issue was found with upstream crash on PowerPC64, where crash
is reporting warning:

    WARNING: cpu 0 invalid NT_PRSTATUS note (n_type != NT_PRSTATUS)
	...

As well as, register values were invalid.

From our analysis, this warning was found in the commit:

	commit db8c030857b4 ("diskdump/netdump: fix segmentation fault caused by failure of stopping CPUs")

The issue arises since with above commit, it becomes mandatory for
crash_notes to be valid for a particular cpu to map percpu prstatus notes.
But some architectures such as PowerPC64, in fadump case
(firmware-assisted dump), does not populate `crash_notes` which is expected
behaviour, since the registers are already stored in the cpu notes in the
vmcore.

In the commit, this changed:

-               if (in_cpu_map(ONLINE_MAP, i)) {
+               if (in_cpu_map(ONLINE_MAP, i) && (!crash_notes_exists || have_crash_notes(i))) {
                        dd->nt_prstatus_percpu[i] = nt_ptr[j++];
                        dd->num_prstatus_notes =
                                MAX(dd->num_prstatus_notes, i+1);

Here, now since the if condition fails, so the `dd->nt_prstatus_percpu` is
not assigned to what was expected, and so register values are also invalid

Signed-off-by: Aditya Gupta <adityag@linux.ibm.com>