diskdump: add hook for additional checks on prstatus notes validity
Introduce a hook in machdep ('is_cpu_prstatus_valid'), for architectures
to determine validity of PRSTATUS notes, when storing percpu PRSTATUS notes

A default hook ('diskdump_is_cpu_prstatus_valid') has also been provided
for all architectures other than PowerPC64, which checks if 'crash_notes'
for a given cpu is valid

PowerPC64 doesn't utilise 'crash_notes' to get register values, so noadditional checks are required

Signed-off-by: Aditya Gupta <adityag@linux.ibm.com>

---

Issue faced due to current code
===============================

Recently an issue was found with upstream crash on PowerPC64, where crash
is reporting warning:

    WARNING: cpu 0 invalid NT_PRSTATUS note (n_type != NT_PRSTATUS)
	...

As well as, register values were invalid.

From our analysis, this warning was found in the commit:

	commit db8c030857b4 ("diskdump/netdump: fix segmentation fault caused by failure of stopping CPUs")

The issue arises since with above commit, it becomes mandatory for
crash_notes to be valid for a particular cpu to map percpu prstatus notes.
But some architectures such as PowerPC64, in fadump case
(firmware-assisted dump), does not populate `crash_notes` which is expected
behaviour, since the registers are already stored in the cpu notes in the
vmcore.

In the commit, this changed:

-               if (in_cpu_map(ONLINE_MAP, i)) {
+               if (in_cpu_map(ONLINE_MAP, i) && (!crash_notes_exists || have_crash_notes(i))) {
                        dd->nt_prstatus_percpu[i] = nt_ptr[j++];
                        dd->num_prstatus_notes =
                                MAX(dd->num_prstatus_notes, i+1);

Here, now since the 'crash_notes' symbol exists, but is not populated, so
the if condition fails, so the `dd->nt_prstatus_percpu` is not assigned
to what was expected, and so register values are also invalid

---