ppc64: do page traversal if vmemmap_list not populated
Currently 'crash-tool' fails on vmcore collected on upstream kernel on
PowerPC64 with the error:

    crash: invalid kernel virtual address: 0  type: "first list entry

Presently the address translation for vmemmap addresses is done using
the vmemmap_list. But with the below commit in Linux, vmemmap_list can
be empty, in case of Radix MMU on PowerPC64

    368a0590d954: (powerpc/book3s64/vmemmap: switch radix to use a
    different vmemmap handling function)

In case vmemmap_list is empty, then it's head is NULL, which crash tries
to access and fails due to accessing NULL.

Instead of depending on 'vmemmap_list' for address translation for
vmemmap addresses, do a kernel pagetable walk to get the physical
address associated with given virtual address

Also, `machdep->machspec->vmemmap_base` can be undefined if
`vmemmap_populate` symbol is not found. Instead of being undefined,
assign `machdep->machspec->vmemmap_base` a default value, similar to how
makedumpfile also does, before checking for the vmemmap symbols

Signed-off-by: Aditya Gupta <adityag@linux.ibm.com>

---

Testing
=======

Git tree with patch applied:
https://github.com/adi-g15-ibm/crash/tree/bugzilla-203296-v4

This can be tested with '/proc/vmcore' as the vmcore, since makedumpfile
also fails in absence of 'vmemmap_list' in upstream linux

The fix for makedumpfile will also been sent to upstream

Changelog
=========

V4
+ elaborate comment, and move code blocks as per reviews on V2 and V3 by
  hari
V3
+ assign `vmemmap_base` to a default value, before vmemmap symbols are
  checked
V2
+ reword commit description and comment to be more clear

---