synchronise cpu context changes between crash/gdb
Currently, both crash context and gdb's current thread context were
pretty independent, and could be different, for example, crash commands
might be working on thread 6 (CPU 5), but GDB passthroughs will be
working on thread 2 (CPU 1).

This was not a problem earlier since interaction of crash and gdb was
not depending on current context for most part. But with more
interactions, this patch should solve any issues due to mismatch in
context in crash/gdb.

Synchronise 'thread' command in gdb with 'set -c' command in crash.
1. crash -> gdb synchronisation:
   Everytime crash's context changes, a helper is called to switch to
   the thread on that CPU in gdb. The function has been implemented in
   crash_target.c, since gdb functions are accessible inside
   'crash_target.c', and the thread ID to CPU ID mapping is also done
   by the crash_target, during initially registering the threads with
   gdb. With this implementation, GDB's default thread initially also
   changes to the crashing thread, so a switch to crashing thread
   manually isn't required anymore

2. gdb -> crash synchronisation:
   gdb has been patched to call 'set_cpu' whenever user switches to any
   thread.

Cc: Sourabh Jain <sourabhjain@linux.ibm.com>
Cc: Hari Bathini <hbathini@linux.ibm.com>
Cc: Mahesh J Salgaonkar <mahesh@linux.ibm.com>
Cc: Naveen N. Rao <naveen.n.rao@linux.vnet.ibm.com>
Cc: Lianbo Jiang <lijiang@redhat.com>
Cc: HAGIO KAZUHITO(萩尾 一仁) <k-hagio-ab@nec.com>
Cc: Tao Liu <ltao@redhat.com>
Signed-off-by: Aditya Gupta <adityag@linux.ibm.com>